+++ CHANGELOG.md
@@ -2,13 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.1.7] - 2025-??-??
+## [1.1.7] - 2025-10-01
 
 ### Changes
 
 - Fixed [issue #10](https://github.com/codedread/spaces/issues/10): Display space window bounds at their last known position and size.
 - Fixed [issue #20](https://github.com/codedread/spaces/issues/20): Close browser windows from the Spaces window.
-- Increased unit test coverage from 10.75% to 16.35%.
+- Fixed [issue #29](https://github.com/codedread/spaces/issues/29): Provide a debug method to export anonymized Spaces DB for debugging.
+- Fixed [issue #30](https://github.com/codedread/spaces/issues/30): Do not prompt twice to overwrite space name.
+- Increased unit test coverage from 10.75% to 16.17%.
 
 ## [1.1.6] - 2025-09-17